refactor(booking): extract duo choice selector to isolated component
