+++ src/features/bookOffer/components/DuoChoiceSelector.tsx
@@ -1,0 +1,59 @@
+import { t } from '@lingui/macro'
+import React from 'react'
+import styled from 'styled-components/native'
+
+import { DuoChoice } from 'features/bookOffer/atoms/DuoChoice'
+import {
+  useBooking,
+  useBookingOffer,
+  useBookingStock,
+} from 'features/bookOffer/pages/BookingOfferWrapper'
+import { useCreditForOffer } from 'features/offer/services/useHasEnoughCredit'
+import { formatToFrenchDecimal } from 'libs/parsers'
+import { Profile } from 'ui/svg/icons/Profile'
+import { IconInterface } from 'ui/svg/icons/types'
+import { getSpacing } from 'ui/theme'
+
+export const DuoChoiceSelector: React.FC = () => {
+  const { bookingState, dispatch } = useBooking()
+  const { isDuo } = useBookingOffer() || {}
+  const stock = useBookingStock()
+  const offerCredit = useCreditForOffer(bookingState.offerId)
+
+  const getChoiceInfosForQuantity = (quantity: 1 | 2) => {
+    const enoughCredit = stock ? quantity * stock.price <= offerCredit : false
+    return {
+      price:
+        enoughCredit && stock
+          ? formatToFrenchDecimal(quantity * stock.price).replace(' ', '')
+          : t`crédit insuffisant`,
+      title: quantity === 1 ? t`Solo` : t`Duo`,
+      selected: bookingState.quantity === quantity,
+      icon: quantity === 1 ? Profile : DuoPerson,
+      onPress: () => dispatch({ type: 'SELECT_QUANTITY', payload: quantity }),
+      hasEnoughCredit: enoughCredit,
+    }
+  }
+
+  return (
+    <DuoChoiceContainer>
+      <DuoChoice {...getChoiceInfosForQuantity(1)} testID={`DuoChoice1`} />
+      {!!isDuo && <DuoChoice {...getChoiceInfosForQuantity(2)} testID={`DuoChoice2`} />}
+    </DuoChoiceContainer>
+  )
+}
+
+const DuoPerson = (props: IconInterface): JSX.Element => (
+  <DuoPersonContainer>
+    <Profile {...props} />
+    <Profile {...props} />
+  </DuoPersonContainer>
+)
+
+const DuoChoiceContainer = styled.View({
+  flexDirection: 'row',
+  flexWrap: 'wrap',
+  marginHorizontal: -getSpacing(2),
+})
+
+const DuoPersonContainer = styled.View({ flexDirection: 'row' })
